requirements/base.txt
aiopg==1.3.4
aioredis==2.0.1
amqp==5.1.1
anyio==3.7.1
asgiref==3.6.0
backports.zoneinfo==0.2.1;python_version<"3.10"
async-timeout==4.0.3
attrs==23.1.0
autobahn==22.4.2
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.5.18
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
daphne==3.0.2
Deprecated==1.2.14
Django==4.2
django-filter==21.1
django-modeltranslation==0.18.2
django-payments==1.0.0
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
django-redis==5.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
exceptiongroup==1.1.3
fastapi==0.100.0
flower==2.0.1
greenlet==2.0.2
h11==0.14.0
humanize==4.8.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.1
MarkupSafe==2.1.1
ocpp==0.19.0
openapi-codec==1.3.2
packaging==21.3
phonenumbers==8.12.48
phonenumberslite==8.13.18
Pillow==9.1.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.12
PyJWT==2.4.0
pymongo==4.2.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytz==2022.1
redis==4.3.4
referencing==0.30.2
requests==2.27.1
rpds-py==0.9.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
sorl-thumbnail==12.8.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
starlette==0.27.0
tornado==6.3.3
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.5.0
uritemplate==4.1.1
uvicorn==0.18.2
vine==5.0.0
wcwidth==0.2.6
websockets==11.0.3
wrapt==1.15.0
zope.interface==5.4.0
django-jazzmin==2.6.0
sentry-sdk==1.32.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-push-notifications==3.0.0
beautifulsoup4==4.12.2
python-environ==0.4.54